[V9990] Update sample program + Fixed compile warning

+++ engine/src/v9990.c
@@ -72,7 +72,7 @@ const u8 g_V9_ModeConfig[] =
 	// Bitmap 1024 * 212   (B7) (undocumented mode)
 	V9_R06_MODE_BITMAP + V9_R06_CLOCK_1 + V9_R06_WIDH_1024 + V9_R06_BPP_4,
 	0,
-	V9_P07_XTAL1,
+	(u8)(V9_P07_XTAL1),
 };
 
 //-----------------------------------------------------------------------------
